@@ -16,15 +16,15 @@
         "animejs": "^3.2.1",
         "bootstrap": "4.3.1",
         "countries-and-timezones": "^3.3.0",
-        "element-ui": "2.15.12",
+        "element-plus": "2.2.25",
         "jquery": "3.5.1",
         "lodash": "4.17.21",
         "moment": "2.29.4",
         "uuid": "^8.3.2",
-        "vue": "2.5.17",
-        "vue-lazyload": "1.3.4",
-        "vue-router": "^3.1.3",
-        "vuex": "3.4.0",
+        "vue": "3.2.45",
+        "vue-lazyload": "3.0.0-rc.2",
+        "vue-router": "4.1.6",
+        "vuex": "4.1.0",
         "workbox-cacheable-response": "^6.1.5",
         "workbox-expiration": "^6.1.5",
         "workbox-routing": "^6.1.5",
@@ -52,10 +52,10 @@
         "eslint-plugin-vue": "^9.8.0",
         "less": "4.1.3",
         "less-loader": "11.0.0",
-        "node-sass": "7.0.1",
-        "postcss": "8.4.14",
+        "node-sass": "8.0.0",
+        "postcss": "8.4.19",
         "prettier": "2.6.2",
-        "sass-loader": "13.0.0",
+        "sass-loader": "13.2.0",
         "typescript": "^4.9.3",
         "vue-template-compiler": "2.5.17",
         "webpack": "5.73.0"
