@@ -40,7 +40,7 @@ import { getRatingColor } from '@/common/utils';
 import moment from 'moment';
 import Vue from 'vue';
 
-export default Vue.extend({
+export default {
     name: 'shareView',
     props: ['configuration', 'showMovieInfo', 'showFullMovieInfo', 'showSeriesInfo', 'movieGenres', 'seriesGenres'],
     data() {
@@ -129,7 +129,7 @@ export default Vue.extend({
             // this.isLoading = false;
         },
     },
-});
+};
 </script>
 
 <style scoped lang="less">
